Add a coursera course for introduction to deep learning
Add "Introduction to Deep Learning" course from coursera. This is a good
course for starting out with fundamental concepts by not going too deep
and not shallow.
For me, this has been one of the best courses for deep learning because it covers all the fundamental concepts in a single course - from ANN to CNN to RNN. Not only that, the assignments are all about the implementation of modern techniques like GAN, Style Transfer, Auto Encoders, etc the knowledge of which is a must to actually understand how modern platforms are leveraging the power of AI.